Begin integration of inventory to interaction/pickup system
I was able to move a portion of the code to where when the player runs over a pickup item, a generic item is now added to inventory (static for now). Still need to pass the pickup data to the inventory system to show accurate results. This is a WIP commit.